storage: implement reviewrs suggestion. Factoring common code between sqlite implementation drivers.
The drivers are now reduced to two functions; one to set the pragmas in the connection uri and the other to return the implementation's registration name.
